Egen fil for reguleringsregler og egen regel for grunnbeløp for å slippe unaturlig grunnlagsobjekt
